Clean up all build scripts.

+++ day4/gold.cc
@@ -20,7 +20,7 @@ std::vector<std::string> split(const std::string& s, char delimiter) {
 
 int get_col_max(const vector<vector<int>> &board, int col) {
     int max = 0;
-    for (int i = 0; i < board.size(); ++i) {
+    for (size_t i = 0; i < board.size(); ++i) {
         auto col_n = board[i][col];
         if (col_n > max)
             max = col_n;
@@ -84,8 +84,8 @@ int main() {
 
     // replace each number with the index of that number in the called-out commands
     for (auto& board : board_turns) {
-        for (int row = 0; row < board.size(); ++row) {
-            for (int col = 0; col < board[0].size(); ++col) {
+        for (size_t row = 0; row < board.size(); ++row) {
+            for (size_t col = 0; col < board[0].size(); ++col) {
                 auto n = board[row][col];
                 auto found = find(begin(cmds), end(cmds), n);
                 if (found != end(cmds))
@@ -100,7 +100,7 @@ int main() {
     // for each row and column of every board, find the largest winning turn for that row/col
     // the smaller of the two, is the winning turn for the whole board
     vector<pair<int, int>> winning_turns; // board index, it's earliest possible winnning turn
-    for (int i = 0; i < board_turns.size(); ++i) {
+    for (size_t i = 0; i < board_turns.size(); ++i) {
         auto board = board_turns[i];
         int min_all_rows = 99999; // smallest maximum of all rows
         for (const auto& row : board) {
@@ -110,7 +110,7 @@ int main() {
         }
 
         int min_all_cols = 99999;
-        for (int col = 0; col < board.size(); ++col) {
+        for (size_t col = 0; col < board.size(); ++col) {
             auto col_max = get_col_max(board, col);
             if (col_max < min_all_cols)
                 min_all_cols = col_max;
